fix: update dtype to f16 for mm

+++ loom/kernels/matmul.py
@@ -27,7 +27,7 @@ def matmul(x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
     assert k == k2
     out = torch.empty([m, n], dtype=torch.promote_types(x.dtype, y.dtype), device=x.device)
     for tile_m, tile_n in hl.tile([m, n]):
-        acc = hl.zeros([tile_m, tile_n], dtype=torch.float32)
+        acc = hl.zeros([tile_m, tile_n], dtype=torch.float16)
         for tile_k in hl.tile(k):
             acc = torch.addmm(acc, x[tile_m, tile_k], y[tile_k, tile_n])
         out[tile_m, tile_n] = acc
@@ -42,8 +42,8 @@ def generate_mlir() -> str:
     Returns:
         Complete MLIR text string (module { module { func.func @matmul ... } }).
     """
-    x = torch.randn([4096, 512], device="cpu", dtype=torch.float32)
-    y = torch.randn([512, 4096], device="cpu", dtype=torch.float32)
+    x = torch.randn([4096, 512], device="cpu", dtype=torch.float16)
+    y = torch.randn([512, 4096], device="cpu", dtype=torch.float16)
     bound_kernel = matmul.bind((x, y))
     print_debug_info(bound_kernel)
     return _helion_generate_mlir(bound_kernel)
